Fix issue where 'shareInstances' resolve but fail to propagate (#201)
* #200 - Expand shareInstances test to replicate propagation issue

* #200 - Fix issue where 'shareInstances' resolve but fail to propagate

Because `$share` is passed by reference to `matchParam()`, and
`matchParam()` removes matching objects from its `$search` array,
instances may be removed from `$share` before it is passed to `expand()`
or `create()`. Depending on the order of constructor parameters and the
relative placement of `shareInstances` dependencies in the object tree,
this may result in multiple instances of these dependencies being
created.

Fixed by passing a copy of the `$share` array to `matchParam()`.